(fix) inputs not setting state properly
-> handleInput had to be refactored, due to the cvData (data.jsx) structure change the function. It sets states depending of a case in the switch statement.
If the setState is education or experience it copies everything up until the children key from the object and sets its value to call the function updateChildrenValues.

(feat) function helper called updateChildrenValues that receives an array, a id and a value and returns a updated version of that received array.
  - more specifically it goes through each element from that array with map(), whenever it reaches the last element it copies the element and changes a value from that object.

(Note) - checking if it works when two groups of inputs were added and the first one is filled first.