(fix) updateChildrenValues helper function
-> this function receives now an extra parameter,
which will be the data-key from the target input,
and will use it within map method to set the input value to the received value.
This ensures the input is placed on the correct group of input fields because it will look for its id.

-> a new variable is set in handleInput that is the event.target.dataset.key
and used it to provide on the newly accepted parameter on the latter helper function.