(refactor) CvRender, & children components.
-> CvRender from the prop form that it receives it sets 3 variables personalInfo, education and experience.
The latter ones are a reference to the array of its group.

-> to reduce code and correct the previous mistake as it would only get the first child,
both RenderEducation & RenderExperience receive one prop only called children(the array got from some of the variables mentioned prior).
Both use map() method to iterate through that array and return its specific information to the elements returned by the component.

Note: the experience information is not showing up. The form is setting experience children with the education children information.